Feat: modelos tipo double

+++ MarianaoWars/Models/Resource.cs
@@ -14,13 +14,13 @@ namespace MarianaoWars.Models
         public double Knowledge { get; set; }
 
         // Indica el almacenamiento de Habilidad invertidos en el ordenador.
-        public int Ingenyous { get; set; }
+        public double Ingenyous { get; set; }
 
         // Indica el almacenamiento de Café invertidos en el ordenador.
-        public int Coffe { get; set; }
+        public double Coffe { get; set; }
 
         // Indica el almacenamiento de Estrés invertidos en el ordenador.
-        public int Stress { get; set; }
+        public double Stress { get; set; }
 
         // Indica el nivel de minas de Conocimiento del ordenador.
         public int KnowledgeLevel { get; set; }
